Move expression vectors from expressions.hp to cre.cpp
